Internal Breeding
Geared to record piglets that are internally bred.  The breed of these pigs is the ID numbers of the sow and sire. This can give insights on how parents affect piglets.

To make recording a lot of piglets more efficient, user can now enter data for multiple piglets at once. This gives rise to the problem of user not know which ID belongs to which pig.